+++ tests/testthat/test-calc_ecoregion.R
@@ -23,4 +23,11 @@ test_that("ecoregion computation works", {
     18352.24,
     tolerance = 1e-4
   )
+
+  # check NA is returned for 0-length tibbles
+  st_geometry(shp) <- st_geometry(shp) + 5
+  st_crs(shp) <- st_crs(4326)
+  expect_equal(
+    .calc_biome(shp, teow),
+    NA)
 })